+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+* Support `.icns` Apple Icon Image format [@toy](https://github.com/toy)
+
 ## v3.5.0 (2026-05-03)
 
 * Add `byte_size` method to expose the size of the image data in bytes [#27](https://github.com/toy/image_size/pull/27) [@dgodd](https://github.com/dgodd)